requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Machine Learning & NLP
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
tokenizers==0.14.1
huggingface-hub>=0.16.4,<0.18
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0

# HTTP & Requests
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
